chore: update ui file naming
