--- example/ui/404.ts
@@ -1,15 +1,0 @@
-const page404: string = String.raw`
-  <!DOCTYPE html>
-  <html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Home</title>
-  </head>
-  <body>
-    <p>Main content</p>
-  </body>
-  </html>
-`
-
-export default page404;
